refactor: shouldLoadModules
