+++ packages/runtime/src/App.tsx
@@ -4,9 +4,8 @@ import type { Navigator } from 'react-router-dom';
 import AppErrorBoundary from './AppErrorBoundary.js';
 import { AppContextProvider } from './AppContext.js';
 import type Runtime from './runtime.js';
-import { createRoutes, matchRoutes } from './routes.js';
+import { createRoutes, shouldLoadModules } from './routes.js';
 import { createTransitionManager } from './transition.js';
-import type { RouteItem, RouteModules } from './types.js';
 
 interface Props {
   runtime: Runtime;
@@ -94,27 +93,4 @@ export default function App(props: Props) {
       </AppErrorBoundary>
     </StrictMode>
   );
-}
-
-/**
- * whether or not execute `transitionManager.handleLoad` function
- */
-function shouldLoadModules(routes: RouteItem[], location: Location, routeModules: RouteModules) {
-  const matches = matchRoutes(routes, location);
-  const loadedModules = matches.filter(match => {
-    const { route: { id } } = match;
-    return routeModules[id];
-  });
-
-  if (loadedModules.length !== matches.length) {
-    // 1. the modules have not been loaded
-    return true;
-  }
-
-  return matches.some((match) => {
-    const { route: { id } } = match;
-    const { getInitialData } = routeModules[id];
-    // 2. if the page route has the getInitialData function, should load page data again
-    return !!getInitialData;
-  });
 }
