+++ packages/runtime/src/routes.tsx
@@ -139,3 +139,26 @@ export function matchRoutes(
     pathnameBase,
   }));
 }
+
+/**
+ * whether or not execute `handleLoad` function in transition manager
+ */
+export function shouldLoadModules(routes: RouteItem[], location: Location, routeModules: RouteModules) {
+  const matches = matchRoutes(routes, location);
+  const loadedModules = matches.filter(match => {
+    const { route: { id } } = match;
+    return routeModules[id];
+  });
+
+  if (loadedModules.length !== matches.length) {
+    // 1. the modules have not been loaded
+    return true;
+  }
+
+  return matches.some((match) => {
+    const { route: { id } } = match;
+    const { getInitialData } = routeModules[id];
+    // 2. if the page route has the getInitialData function, should load page data again
+    return !!getInitialData;
+  });
+}
